+++ my_ptr_put_nbr.c
@@ -2,5 +2,5 @@
 
 void	my_ptr_putnbr(va_list ap)
 {
-	my_put_nbr(va_arg(ap, int));
+	my_put_nbr(va_arg(ap, long int));
 }
